requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-dotenv==1.1.0
notion-client==2.3.0
httpx==0.28.1
pyyaml==6.0.2
requests
